added bookmarks as its own component, issue with memory leak on bookmark/reference
